movie-agent/pyproject.toml
[project]
name = "movie-agent"
version = "0.1.0"
description = "Movie Expert Agent - OpenAI Agents SDK + Clean Architecture"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "httpx>=0.28.0",
    "ipykernel>=7.2.0",
]
